notification-service/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
psycopg[binary]==3.2.12
gunicorn==23.0.0
Pillow>=12.0.0
channels==4.0.0
channels-redis==4.2.1
daphne==4.1.2
requests==2.32.3
pika==1.3.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
faker==20.1.0
